# Simplify to_pydatetime()