--- a/pandas/_libs/tslib.pyx
+++ b/pandas/_libs/tslib.pyx
@@ -1162,8 +1162,8 @@ cdef class _Timestamp(datetime):
                           UserWarning, stacklevel=2)
 
         return datetime(self.year, self.month, self.day,
-                         self.hour, self.minute, self.second,
-                         self.microsecond, self.tzinfo)
+                        self.hour, self.minute, self.second,
+                        self.microsecond, self.tzinfo)
 
     cpdef to_datetime64(self):
         """ Returns a numpy.datetime64 object with 'ns' precision """