@pep8speaks 2017-09-20
Hello @jbrockmendel! Thanks for updating the PR.

 - In the file [`asv_bench/benchmarks/timestamp.py`](https://github.com/pandas-dev/pandas/blob/ee2885911935238c4a369162aedb50204a612e97/asv_bench/benchmarks/timestamp.py), following are the PEP8 issues :

> [Line 90:1](https://github.com/pandas-dev/pandas/blob/ee2885911935238c4a369162aedb50204a612e97/asv_bench/benchmarks/timestamp.py#L90): [W391](https://duckduckgo.com/?q=pep8%20W391) blank line at end of file