@jbrockmendel 2017-09-20
Looks like someone added one overnight, will add to_pydatetime timing to it.