@jbrockmendel 2017-09-21
> looks fine
>
>you can click thru to circleci to see errors

I don't ask _just_ because I want to waste your time.  I'm saying I don't see any error message in the log.